main pull

+++ rest_app/views/user_view.py
@@ -1,10 +1,9 @@
 from flask import Blueprint, render_template, redirect, url_for, flash
 from flask_login import login_required
-from rest_app.service.user_service import form_user_data_parser, update_user, add_user
-from rest_app.service.address_service import add_address
+from rest_app.service.user_service import form_user_data_parser, update_user, add_user, get_all_users_by_is_admin_filter
+from rest_app.service.address_service import add_address, address_data_form_parser
 from rest_app.service.common_services import delete_row_by_id
-from rest_app.forms.personal_info_forms import UpdateProfileForm, AddressForm, \
-    add_values_to_address_form, add_values_to_profile_form
+from rest_app.forms.personal_info_forms import UpdateProfileForm, AddressForm
 from rest_app.models import Address, User
 from rest_app.forms.admin_forms import AddUser
 from rest_app.views.admin_views import admin_or_user_required, admin_required
@@ -12,12 +11,39 @@ from rest_app.views.admin_views import admin_or_user_required, admin_required
 user = Blueprint('user', __name__, url_prefix='/user')
 
 
-@user.route('/<string:user_id>')
-@admin_or_user_required
+@user.route('/<string:user_id>', methods=['GET', 'POST'])
 @login_required
+@admin_or_user_required
 def user_detail(user_id):
-    address_form = add_values_to_address_form(AddressForm(), Address, user_id)
-    profile_form = add_values_to_profile_form(UpdateProfileForm(user_id), User, user_id)
+    address_form = AddressForm()
+    profile_form = UpdateProfileForm(user_id)
+    populate_personal_info(address_form, profile_form, user_id)
+
+    return render_template(
+        'personal_info.html',
+        profile_form=profile_form,
+        address_form=address_form,
+        user_id=user_id
+    )
+
+
+@user.route('/update_user/<string:user_id>', methods=['POST'])
+def update_user_details(user_id):
+    address_form = AddressForm()
+    profile_form = UpdateProfileForm(user_id)
+
+    if profile_form.submit_profile.data and profile_form.validate():
+        args = form_user_data_parser().parse_args()
+        update_user(user_id, **args)
+        flash('User information was updated', 'success')
+        return redirect(url_for('user.user_detail', user_id=user_id))
+    elif address_form.submit_address.data and address_form.validate():
+        args = address_data_form_parser().parse_args()
+        add_address(user_id=user_id, **args)
+        flash('Address information has been updated', 'success')
+        return redirect(url_for('user.user_detail', user_id=user_id))
+
+    populate_personal_info(address_form, profile_form, user_id)
 
     return render_template(
         'personal_info.html',
@@ -28,8 +54,8 @@ def user_detail(user_id):
 
 
 @user.route('/create', methods=['GET', 'POST'])
-@admin_required
 @login_required
+@admin_required
 def create_user():
     form = AddUser()
 
@@ -39,54 +65,36 @@ def create_user():
         flash('User was added successfully', 'success')
         return redirect(url_for('admin.admin_main'))
 
-    form.phone_number.data = '+1'
     return render_template('add_user.html', form=form)
 
 
-@user.route('/<string:user_id>/update', methods=['POST'])
-@admin_or_user_required
-@login_required
-def update_user_values(user_id):
-    profile_form = UpdateProfileForm(user_id)
-    address_form = AddressForm()
-
-    if profile_form.validate_on_submit():
-        args = form_user_data_parser().parse_args()
-        update_user(user_id, **args)
-        flash('User information was updated', 'success')
-        return redirect(url_for('user.user_detail', user_id=user_id))
-
-    address_form = add_values_to_address_form(address_form, Address, user_id)
-    profile_form = add_values_to_profile_form(profile_form, User, user_id)
-    return render_template('personal_info.html', profile_form=profile_form, address_form=address_form)
-
-
 @user.route('/<string:user_id>/delete')
-@admin_required
 @login_required
+@admin_required
 def delete_user(user_id):
-    delete_row_by_id(User, user_id)
-    flash('User was successfully deleted', 'success')
+    user = User.query.filter_by(id=user_id).first()
+
+    if not user.is_admin:
+        delete_row_by_id(User, user_id)
+        flash('User was successfully deleted', 'success')
+    else:
+        if len(get_all_users_by_is_admin_filter(True).all()) > 1:
+            flash('You are the only admin. Please create a new one and after repeat', 'danger')
+
     return redirect(url_for('admin.admin_main'))
 
 
-@user.route('/<string:user_id>/update_address', methods=['POST'])
-@admin_or_user_required
-@login_required
-def update_address_details(user_id):
-    profile_form = UpdateProfileForm(user_id)
-    address_form = AddressForm()
+def populate_personal_info(address_form, profile_form, user_id):
+    """
+    Pre-populates values in form fields
 
-    if address_form.validate_on_submit():
-        add_address(
-            user_id=user_id,
-            city=address_form.city.data,
-            postal_code=address_form.postal_code.data,
-            street=address_form.street.data,
-            street_number=address_form.street_number.data
-        )
-        flash('Address information has been updated', 'success')
-        return redirect(url_for('user.user_detail', user_id=user_id))
+    :param address_form: instance of address from
+    :param profile_form: instance of profile from
+    :param user_id: unique id of the user
+    """
+    addresses = Address.query.filter(Address.user_id == user_id).all()
 
-    profile_form, address_form = populate_form_values(address_form, profile_form, user_id, Address, User)
-    return render_template('personal_info.html', profile_form=profile_form, address_form=address_form)
+    if addresses:
+        address_form.change_address_values(addresses.pop())
+
+    profile_form.change_user_form_values(user_id)
